HSEARCH-3638 Reorganize multi-tenancy tests to have the same base configuration in all methods
Otherwise we won't be able to apply the changes of the next commit.
