+++ integrationtest/backend/tck/src/main/java/org/hibernate/search/integrationtest/backend/tck/multitenancy/MultiTenancyBaseIT.java
@@ -4,7 +4,7 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.search.integrationtest.backend.tck;
+package org.hibernate.search.integrationtest.backend.tck.multitenancy;
 
 import static org.hibernate.search.util.impl.integrationtest.common.assertion.SearchResultAssert.assertThat;
 import static org.hibernate.search.util.impl.integrationtest.common.stub.mapper.StubMapperUtils.referenceProvider;
@@ -33,7 +33,7 @@ import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.ExpectedException;
 
-public class MultiTenancyIT {
+public class MultiTenancyBaseIT {
 
 	public static final String CONFIGURATION_ID = "multi-tenancy";
 
@@ -331,122 +331,6 @@ public class MultiTenancyIT {
 	}
 
 	@Test
-	public void backend_multi_tenancy_disabled_but_indexes_requiring_multi_tenancy_throws_exception() {
-		thrown.expect( SearchException.class );
-		thrown.expectMessage( "Index" );
-		thrown.expectMessage( "requires multi-tenancy but the backend" );
-		thrown.expectMessage( "does not support it in its current configuration." );
-
-		setupHelper.withDefaultConfiguration()
-				.withIndex(
-						INDEX_NAME,
-						ctx -> this.indexMapping = new IndexMapping( ctx.getSchemaElement() ),
-						indexManager -> this.indexManager = indexManager
-				)
-				.withMultiTenancy()
-				.setup();
-
-		initData();
-	}
-
-	@Test
-	public void using_multi_tenancy_for_query_while_disabled_throws_exception() {
-		thrown.expect( SearchException.class );
-		thrown.expectMessage( "Tenant identifier" );
-		thrown.expectMessage( "is provided, but multi-tenancy is disabled for this backend" );
-
-		setupHelper.withDefaultConfiguration()
-				.withIndex(
-						"IndexName-using_multi_tenancy_for_query_while_disabled_throws_exception",
-						ctx -> this.indexMapping = new IndexMapping( ctx.getSchemaElement() ),
-						indexManager -> this.indexManager = indexManager
-				)
-				.setup();
-
-		StubMappingScope scope = indexManager.createScope();
-		SearchQuery<DocumentReference> query = scope.query( tenant1SessionContext )
-				.predicate( f -> f.matchAll() )
-				.toQuery();
-		assertThat( query )
-				.hasDocRefHitsAnyOrder( INDEX_NAME, DOCUMENT_ID_1, DOCUMENT_ID_2 );
-	}
-
-	@Test
-	public void using_multi_tenancy_for_add_while_disabled_throws_exception() {
-		thrown.expect( SearchException.class );
-		thrown.expectMessage( "Tenant identifier" );
-		thrown.expectMessage( "is provided, but multi-tenancy is disabled for this backend" );
-
-		setupHelper.withDefaultConfiguration()
-				.withIndex(
-						"IndexName-using_multi_tenancy_for_add_while_disabled_throws_exception",
-						ctx -> this.indexMapping = new IndexMapping( ctx.getSchemaElement() ),
-						indexManager -> this.indexManager = indexManager
-				)
-				.setup();
-
-		IndexWorkPlan<? extends DocumentElement> workPlan = indexManager.createWorkPlan( tenant1SessionContext );
-
-		workPlan.add( referenceProvider( DOCUMENT_ID_3 ), document -> {
-			document.addValue( indexMapping.string, STRING_VALUE_3 );
-			document.addValue( indexMapping.integer, INTEGER_VALUE_5 );
-
-			DocumentElement nestedObject = document.addObject( indexMapping.nestedObject.self );
-			nestedObject.addValue( indexMapping.nestedObject.string, STRING_VALUE_3 );
-			nestedObject.addValue( indexMapping.nestedObject.integer, INTEGER_VALUE_5 );
-		} );
-
-		workPlan.execute().join();
-	}
-
-	@Test
-	public void using_multi_tenancy_for_update_while_disabled_throws_exception() {
-		thrown.expect( SearchException.class );
-		thrown.expectMessage( "Tenant identifier" );
-		thrown.expectMessage( "is provided, but multi-tenancy is disabled for this backend" );
-
-		setupHelper.withDefaultConfiguration()
-				.withIndex(
-						"IndexName-using_multi_tenancy_for_update_while_disabled_throws_exception",
-						ctx -> this.indexMapping = new IndexMapping( ctx.getSchemaElement() ),
-						indexManager -> this.indexManager = indexManager
-				)
-				.setup();
-
-		IndexWorkPlan<? extends DocumentElement> workPlan = indexManager.createWorkPlan( tenant1SessionContext );
-
-		workPlan.update( referenceProvider( DOCUMENT_ID_2 ), document -> {
-			document.addValue( indexMapping.string, UPDATED_STRING );
-			document.addValue( indexMapping.integer, INTEGER_VALUE_4 );
-
-			DocumentElement nestedObject = document.addObject( indexMapping.nestedObject.self );
-			nestedObject.addValue( indexMapping.nestedObject.string, UPDATED_STRING );
-			nestedObject.addValue( indexMapping.nestedObject.integer, INTEGER_VALUE_4 );
-		} );
-
-		workPlan.execute().join();
-	}
-
-	@Test
-	public void using_multi_tenancy_for_delete_while_disabled_throws_exception() {
-		thrown.expect( SearchException.class );
-		thrown.expectMessage( "Tenant identifier" );
-		thrown.expectMessage( "is provided, but multi-tenancy is disabled for this backend" );
-
-		setupHelper.withDefaultConfiguration()
-				.withIndex(
-						"IndexName-using_multi_tenancy_for_delete_while_disabled_throws_exception",
-						ctx -> this.indexMapping = new IndexMapping( ctx.getSchemaElement() ),
-						indexManager -> this.indexManager = indexManager
-				)
-				.setup();
-
-		IndexWorkPlan<? extends DocumentElement> workPlan = indexManager.createWorkPlan( tenant1SessionContext );
-		workPlan.delete( referenceProvider( DOCUMENT_ID_1 ) );
-		workPlan.execute().join();
-	}
-
-	@Test
 	public void not_using_multi_tenancy_for_query_while_enabled_throws_exception() {
 		thrown.expect( SearchException.class );
 		thrown.expectMessage( "Backend" );
@@ -456,8 +340,6 @@ public class MultiTenancyIT {
 		SearchQuery<DocumentReference> query = scope.query( new StubSessionContext() )
 				.predicate( f -> f.matchAll() )
 				.toQuery();
-		assertThat( query )
-				.hasDocRefHitsAnyOrder( INDEX_NAME, DOCUMENT_ID_1, DOCUMENT_ID_2 );
 	}
 
 	@Test
